vcg::tri::Allocate<MESH>::Get*Attribute(m,name);  has CHANGED!
Before it was returning a valid handle to the attribute
IF it existed, otherwise the handle was invalid.
 
Now it always returns a valid handle to the attribute,
by creating the attribute name if it did not exist or returning
the handle to it if it was already there.

ALl the code using Get*Attribute  has been changed accordingly.





git-svn-id: svn://svn.code.sf.net/p/vcg/code/trunk/vcglib@4811 e8973e21-934d-0410-ac44-de4585f5606d